+++ src/Extension/Build.php
@@ -8,6 +8,17 @@ use Version\Assert\VersionAssert;
 
 class Build extends BaseExtension
 {
+    public static function empty(): Build
+    {
+        static $noBuild = null;
+
+        if (null === $noBuild) {
+            $noBuild = new self(...[]);
+        }
+
+        return $noBuild;
+    }
+
     protected function validate(string $identifier): void
     {
         VersionAssert::that($identifier)->regex(